Moves /pkg/internal/apis/istio to /pkg/internal/istio
