+++ pkg/controller/acmechallenges/BUILD.bazel
@@ -21,7 +21,7 @@ go_library(
         "//pkg/controller:go_default_library",
         "//pkg/controller/acmechallenges/scheduler:go_default_library",
         "//pkg/feature:go_default_library",
-        "//pkg/internal/apis/istio:go_default_library",
+        "//pkg/internal/istio:go_default_library",
         "//pkg/issuer:go_default_library",
         "//pkg/issuer/acme/dns:go_default_library",
         "//pkg/issuer/acme/dns/util:go_default_library",
