+++ pkg/internal/BUILD.bazel
@@ -13,8 +13,8 @@ filegroup(
         "//pkg/internal/api/validation:all-srcs",
         "//pkg/internal/apis/acme:all-srcs",
         "//pkg/internal/apis/certmanager:all-srcs",
-        "//pkg/internal/apis/istio:all-srcs",
         "//pkg/internal/apis/meta:all-srcs",
+        "//pkg/internal/istio:all-srcs",
         "//pkg/internal/vault:all-srcs",
     ],
     tags = ["automanaged"],
